+++ server/lib/elastic-wrapper.js
@@ -14,6 +14,7 @@ import { monitoringKnownFields } from '../integration-files/monitoring-known-fie
 
 export class ElasticWrapper {
   constructor(server) {
+    this.cachedCredentials = {};
     this._server = server;
     this.usingSearchGuard = ((server || {}).plugins || {}).searchguard || false;
     this.elasticRequest = server.plugins.elasticsearch.getCluster('data');
@@ -505,6 +506,14 @@ export class ElasticWrapper {
     }
   }
 
+  cleanCachedCredentials() {
+    const now = new Date().getTime();
+    for (const key in this.cachedCredentials) {
+      if (now - this.cachedCredentials[key].fetched_at >= 2000)
+        delete this.cachedCredentials[key];
+    }
+  }
+
   /**
    * Search for the Wazuh API configuration document using its own id (usually it's a timestamp)
    * @param {*} id Eg: 12396176723
@@ -512,24 +521,31 @@ export class ElasticWrapper {
   async getWazuhConfigurationById(id) {
     try {
       if (!id) return Promise.reject(new Error('No valid document id given'));
+      this.cleanCachedCredentials();
+      if (!this.cachedCredentials[id]) {
+        const data = await this.elasticRequest.callWithInternalUser('get', {
+          index: '.wazuh',
+          type: '_doc',
+          id: id
+        });
+        const object = {
+          user: data._source.api_user,
+          secret: Buffer.from(data._source.api_password, 'base64').toString(
+            'ascii'
+          ),
+          url: data._source.url,
+          port: data._source.api_port,
+          insecure: data._source.insecure,
+          cluster_info: data._source.cluster_info,
+          extensions: data._source.extensions,
+          fetched_at: new Date().getTime()
+        };
+        this.cachedCredentials[id] = { ...object };
+      }
 
-      const data = await this.elasticRequest.callWithInternalUser('get', {
-        index: '.wazuh',
-        type: '_doc',
-        id: id
-      });
+      this.cachedCredentials[id].password = this.cachedCredentials[id].secret;
 
-      return {
-        user: data._source.api_user,
-        password: Buffer.from(data._source.api_password, 'base64').toString(
-          'ascii'
-        ),
-        url: data._source.url,
-        port: data._source.api_port,
-        insecure: data._source.insecure,
-        cluster_info: data._source.cluster_info,
-        extensions: data._source.extensions
-      };
+      return this.cachedCredentials[id];
     } catch (error) {
       return Promise.reject(error);
     }